Project Files/Django Website/requirements.txt
astroid==2.3.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.1
Django==2.2.7
django-crispy-forms==1.11.2
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.2.0
PillowImage==1.1.11
PyBundle==1.0.6
pylint==2.4.4
pytz==2019.3
reportlab==3.5.67
requests==2.25.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.26.4
wrapt==1.11.2
